deployment-engine/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-jose==3.3.0
python-multipart==0.0.6
requests==2.30.0
httpx==0.24.0
sqlalchemy==2.0.12
psycopg2-binary==2.9.6
python-dotenv==1.0.0
boto3==1.26.133
azure-identity==1.13.0
azure-mgmt-resource==22.0.0
google-cloud-resource-manager==1.10.1
terraform-wrapper==0.1.1
pyyaml==6.0
jinja2==3.1.2

